un-f*ck the JSON project import feature
Basically, thanks to me mostly just vibe coding this dumb project, there was a time where the the property name for indicating line breaks and new subtitles (cards) was isNewLine and isNewCard. But when I needed to change some behavior, it made sense to change the names to newLineAfter and newCardAfter.

This broke the import for proj json files that were made and saved before this change was made.

The fix was to define the state on proj json file load to use the newLineAfter and newCardAfter properties by reading both the current version and the old version, and just to be clear, remove any future exports

Because the proj.json files didn't import right, I assumed I hadn't split a number of my own transcripts, as a result, I wasted nearly 2 hours of my time re-splitting transcripts that I'd already done two weeks ago and erased my previous work in the process.

In addition, I have further project files that appear to be completely blank, and I have no idea why. Did I load them and resave them to just make them immediately blank?? I am at a total loss

I should've never updated the property name, it was so f*cking pointless to do so, and has cost me hours of wasted time. I am furious.